polished sidebar and cleaned up css
