+++ src/App.css
@@ -1,15 +1,19 @@
-/* .App {
-  text-align: center;
+.background {
+  display: flex;
+  background-color: #A8D0E6;
+  min-height: 100vh;
 }
 
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+.content-pane {
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+  margin: auto;
 }
-*/
+
+.divider {
+  padding: 2rem 0;
+}
+
+.divider:not(:first-child) {
+  border-top: 5px solid #374785;
+}
