+++ src/Pages/index.js
@@ -1,4 +1,3 @@
 export { AboutMe } from './AboutMe'
 export { ContactMe } from './ContactMe'
 export { MyExperiences } from './MyExperiences'
-export { HomePage } from './HomePage'
